--- fourtoon-cookie/src/pages/SignUpPage/BirthInputLayout/BirthInputLayout.styled.ts
@@ -1,14 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    alignContent: 'center',
-    backgroundColor: '#f7f7f7',
-    borderRadius: 12,
-    width: 350,
-    height: 280,
-  },
-  });
-
