--- fourtoon-cookie/src/pages/SignUpPage/GenderInputLayout/GenderInputLayout.styled.ts
@@ -1,30 +1,0 @@
-import { StyleSheet } from "react-native";
-
-
-export const styles = StyleSheet.create({
-    genderOption: {
-      marginTop: '60%',
-      width: 147,
-      height: 138,
-      padding: 10,
-      borderColor: '#EEEEEE',
-      borderWidth: 1,
-      borderRadius: 16,
-      justifyContent: 'center',
-      alignItems: 'center',
-    },
-    selectedOption: {
-      borderColor: '#FFC426',
-    },
-    genderImage: {
-      width: 60,
-      height: 60,
-      marginBottom: 10,
-    },
-    genderText: {
-      fontSize: 17,
-      fontWeight: '600',
-      color: '#212121',
-      marginTop: 10,
-    },
-  });
