--- fourtoon-cookie/src/pages/SignUpPage/Header/Header.styled.ts
@@ -1,22 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    header: {
-      flexDirection: 'column',
-      alignItems: 'flex-start',
-      marginBottom: 10,
-    },
-    headerText: {
-      fontSize: 16,
-      fontWeight: '500',
-      color: '#AAAAAA',
-      marginTop: 20,
-    },
-    backButton: {
-      width: 24,
-      height: 24,
-      justifyContent: 'center',
-      alignItems: 'flex-end',
-      position: 'relative'
-    }
-  });
