--- fourtoon-cookie/src/pages/SignUpPage/NameInputLayout/NameInputLayout.styled.ts
@@ -1,14 +1,0 @@
-import { StyleSheet } from "react-native";
-
-
-export const styles = StyleSheet.create({
-    input: {
-      borderBottomWidth: 1,
-      borderBottomColor: '#212121',
-      fontSize: 24,
-      width: '100%',
-      paddingVertical: 5,
-    },
-  });
-
-
