+++ fourtoon-cookie/src/pages/SignUpPage/SignUpPage.tsx
@@ -1,18 +1,17 @@
 import { ReactNode, useEffect, useState } from "react";
 import { Gender } from "../../types/gender";
 import ProgressBar from "../../components/common/ProgressBar/ProgressBar";
-import { Keyboard, KeyboardAvoidingView, SafeAreaView, Text, View } from "react-native";
-import * as S from "./SignUpPage.styled";
-import Header from "./Header/Header";
+import { Keyboard, KeyboardAvoidingView, SafeAreaView, StyleSheet, Text, View } from "react-native";
+import Header from "./Header";
 import Button from "../../components/common/Button/Button";
-import NameInputLayout from "./NameInputLayout/NameInputLayout";
-import BirthInputLayout from "./BirthInputLayout/BirthInputLayout";
-import GenderInputLayout from "./GenderInputLayout/GenderInputLayout";
+import NameInputLayout from "./NameInputLayout";
+import BirthInputLayout from "./BirthInputLayout";
+import GenderInputLayout from "./GenderInputLayout";
 import { NavigationProp, useNavigation } from "@react-navigation/native";
 import { RootStackParamList } from "../../types/routing";
 import { LocalDate, use } from "@js-joda/core";
 
-import AgreementInputLayout from "./AgreementInputLayout/AgreementInputLayout";
+import AgreementInputLayout from "./AgreementInputLayout";
 import { useAccountState } from "../../hooks/account";
 import { useFunctionWithErrorHandling } from "../../hooks/error";
 import { useTranslationWithParentName } from "../../hooks/locale";
@@ -93,8 +92,8 @@ const SignUpPage = () => {
     });
 
     return (
-        <SafeAreaView style={S.styles.safeArea}>
-            <View style={S.styles.container}>
+        <SafeAreaView style={styles.safeArea}>
+            <View style={styles.container}>
                 <Header onBackButtonPress={handleBackButtonPress}/>
                 {
                     signUpProgress == SignUpProgres.NAME && 
@@ -132,12 +131,12 @@ const SignUpPage = () => {
                     </Container>
                 }
                 <KeyboardAvoidingView 
-                    style={S.styles.bottomContainer} 
+                    style={styles.bottomContainer} 
                     enabled={true}
                     keyboardVerticalOffset={80}
                     behavior={'padding'}
                 >
-                    <View style={S.styles.progressContainer}>
+                    <View style={styles.progressContainer}>
                         <ProgressBar
                             progress={signUpProgress}
                             totalProgress={4}
@@ -148,10 +147,10 @@ const SignUpPage = () => {
                         title={commonT('next')}
                         onPress={handleNextButtonClick}
                         style={{
-                            ...S.styles.nextButton, 
+                            ...styles.nextButton, 
                             backgroundColor: isNextButtonAvailabe ? '#FFC426' : '#DDDDDD'
                         }}
-                        textStyle={S.styles.nextButtonText}
+                        textStyle={styles.nextButtonText}
                     />
                 </KeyboardAvoidingView>
             </View>
@@ -172,10 +171,53 @@ const Container = (props: ContainerProps) => {
 
     return (
         <View>
-            <Text style={S.styles.title}>{title}</Text>
-            <View style={S.styles.inputContainer}>
+            <Text style={styles.title}>{title}</Text>
+            <View style={styles.inputContainer}>
                 {children}
             </View>
         </View>
     );
 }
+
+const styles = StyleSheet.create({
+    safeArea: {
+      flex: 1,
+      backgroundColor: '#FFFFFF',
+    },
+    container: {
+      flex: 1,
+      padding: 23,
+      position: 'relative'
+    },
+    title: {
+      fontSize: 28,
+      fontWeight: '700',
+      color: '#212121'
+    },
+    inputContainer: {
+      flexDirection: 'row',
+      justifyContent: 'space-around',
+      marginTop: 50,
+      marginBottom: 20,
+    },
+    bottomContainer: {
+      position: 'absolute',
+      bottom: 20,
+      left: 20,
+      right: 20,
+    },
+    progressContainer: {
+      marginBottom: 20,
+    },
+    nextButton: {
+      width: '100%',
+      height: 60,
+      borderRadius: 16,
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    nextButtonText: {
+      fontSize: 17,
+      fontWeight: '600'
+    }
+});
